+++ unit_tests/reduct/storage/entry_query_test.cc
@@ -4,6 +4,7 @@
 #include <google/protobuf/util/time_util.h>
 
 #include <filesystem>
+#include <thread>
 
 #include "reduct/helpers.h"
 #include "reduct/storage/entry.h"
