+++ unit_tests/reduct/storage/storage_entry_api_test.cc
@@ -8,7 +8,6 @@
 #include "reduct/helpers.h"
 #include "reduct/storage/storage.h"
 
-using reduct::async::Run;
 using reduct::async::Task;
 using reduct::core::Error;
 
@@ -16,8 +15,7 @@ using reduct::proto::api::BucketSettings;
 using reduct::storage::IStorage;
 
 using reduct::MakeDefaultBucketSettings;
-using reduct::OnNextRecord;
-using reduct::OnQuery;
+
 
 using reduct::async::IAsyncReader;
