+++ unit_tests/reduct/storage/storage_server_api_test.cc
@@ -5,7 +5,6 @@
 #include "reduct/helpers.h"
 #include "reduct/storage/storage.h"
 
-using reduct::async::Task;
 using reduct::core::Error;
 
 using reduct::storage::IStorage;
